Move query to phenotype search
Issue: IBP-3182
Reviewer: Clary
